Create pseudo-saveYDK
This is just my design for the save functionality of the deck editor.
The saveFileRequest is obviously not done, but the handleSaveFileRequest
will work, assuming it is given the object with deck arrays and the deck
name as params. Committing to a separate file because I want it to be
looked at first.