add missing strings
I copied the invisible character between the words , "Normal" and "Summon" in `!system 1 Normal Summon` , so it shouldnt cut off in Salvation :

`!system 66 持续公开`
Since we dont know what it is yet , im still looking , and if it shows up as that the only Chinese string , then we should know what its for
Google Translate translates it as "Continued public" , so i assume its used in situations where you want to continue to keep cards revealed for whatever reason..
`!system 569 Select a Zone to play the card`
`!system 570 Select a Zone to become unusable`
https://github.com/SalvationDevelopment/YGOPro-Salvation-Server/pull/308#issuecomment-276984394
Zone instead of Zone(s) cus Ygopro shows you the min./max. numbers in brackets after the string in Ygopro ,
For eg. "Select a Zone to become unusable (1/2)"